redundancy and bug fix
Added a second relay that turns on for 
all the duration of the burnout cycle
and functions as a redundant emergency
shutdown in case either relay fails.

also fixed a bug in regard to bignumbers,
I forgot to create the chars, they
seem to have been working because they
remained in the memory of the lcd or
something like that.
but for some reason the liquidCrystal library
will not allow me to create 5 chars, even
though the documentation says you can create
up to 8, it crashed the SD library whenever
I tried adding a 5th character.
so I replaced the degree symbol with an "o".
to be honest it works better this way.